Clean up map's interactive targets when tiles are removed.
Currently, all features ever created for any tile amass in the map's
interactive targets, preventing garbage collection.

My solution only works when VectorGrid is recording created tiles (i.e.
with getFeatureId), but at least it's better than nothing.

It is worth considering to record created tiles when interactive in order
to be able to do this irrespective of the getFeatureId option.