Make features instead of tiles accept pointer events.
If the layer is interactive, making the tiles accept pointer events
prevents makes it impossible to "click through" them if used as overlays.
Making the features accept pointer events instead makes overlays work as
expected.

Note that the path.leaflet-interactive CSS rule does not apply because it
is limited to SVGs that are direct descendants of a .leaflet-pane, while
our SVGs live many layers down inside a .leaflet-layer. Changing the CSS
rule to apply to all SVG <path>s marked .leaflet-interactive underneath
.leaflet-pane would avoid the need to manually set the 'pointer-events'
style on each <path>, but might have unintended effects which need to be
investigated.